Finish T0X.01-Exercise-GoogleApiClient.
Can request fine location permission and connect to GoogleApiClient.
The code commented out in MainActivity is my initial version, which
can show an alert dialog and requests permission again, and
also still works on both API 17 tablet and API 28 emulator, but isn't
as clean. The layout and checkbox XML commented out in activity_main does the
same thing as the existing checkbox XML.